requirements.txt
asttokens==2.4.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
# pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.9.0
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0

transformers
